back-end/requirements.txt
black==21.9b0
coverage==6.0.1
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Restless-NG==2.2.9
Flask-SQLAlchemy==2.5.1
flask-marshmallow==0.14.0
flask_restless
gunicorn==20.1.0
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mypy==0.910
mypy-extensions==0.4.3
pylint==2.11.1
pylint-flask-sqlalchemy==0.2.0
python-dotenv==0.20.0
regex==2021.10.8
requests==2.27.1
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.2
uWSGI==2.0.20
xmltodict==0.12.0
psycopg2-binary
pytest==7.0.1
